pyproject.toml
[project]
name = "MCP-NAME"
version = "0.0.0"
description = "MCP Server Template with logging and transport configuration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
mcp-server = "src.MCP_NAME.mcp_main:main"
